# docs/api/v1-openapi.json
{
  "components": {
    "schemas": {
      "AssessmentResult": {
        "description": "The recommended tier and an indicative monthly price.",
        "properties": {
          "assessment_id": {
            "title": "Assessment Id",
            "type": "string"
          },
          "indicative_monthly": {
            "$ref": "#/components/schemas/Money"
          },
          "recommended_tier": {
            "title": "Recommended Tier",
            "type": "string"
          },
          "score": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "integer"
          }
        },
        "required": [
          "assessment_id",
          "score",
          "recommended_tier",
          "indicative_monthly"
        ],
        "title": "AssessmentResult",
        "type": "object"
      },
      "AssessmentSubmit": {
        "additionalProperties": false,
        "description": "A completed funnel submission; contact data is minimised (DPIA-001).",
        "properties": {
          "congregation_size": {
            "maximum": 1000000.0,
            "minimum": 0.0,
            "title": "Congregation Size",
            "type": "integer"
          },
          "contact_email": {
            "anyOf": [
              {
                "maxLength": 320,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "locale": {
            "default": "lt",
            "pattern": "^(lt|en|ru|pl|de)$",
            "title": "Locale",
            "type": "string"
          },
          "parish": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Parish",
            "type": "string"
          },
          "utm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UtmParams"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "parish",
          "congregation_size"
        ],
        "title": "AssessmentSubmit",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Money": {
        "description": "An amount in integer minor units plus an ISO-4217 currency code.\n\nFloats are never used for money (ADR-006); EUR is the default and the only\ncurrency currently supported by the pricing engine.",
        "properties": {
          "amount_minor": {
            "description": "Amount in minor units (euro cents).",
            "title": "Amount Minor",
            "type": "integer"
          },
          "currency": {
            "default": "EUR",
            "pattern": "^[A-Z]{3}$",
            "title": "Currency",
            "type": "string"
          }
        },
        "required": [
          "amount_minor"
        ],
        "title": "Money",
        "type": "object"
      },
...
      "QuoteLine": {
        "description": "One priced line on a quote.",
        "properties": {
          "description": {
            "title": "Description",
            "type": "string"
          },
          "quantity": {
            "minimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          },
          "recurring": {
            "default": false,
            "title": "Recurring",
            "type": "boolean"
          },
          "total": {
            "$ref": "#/components/schemas/Money"
          },
          "unit": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "description",
          "unit",
          "quantity",
          "total"
        ],
        "title": "QuoteLine",
        "type": "object"
      },
      "QuoteRequest": {
        "additionalProperties": false,
        "description": "A pricing request for one tier with optional extras.",
        "properties": {
          "annual_prepay": {
            "default": false,
            "title": "Annual Prepay",
            "type": "boolean"
          },
          "extra_seats": {
            "default": 0,
            "maximum": 1000.0,
            "minimum": 0.0,
            "title": "Extra Seats",
            "type": "integer"
          },
          "tier": {
            "pattern": "^(essential|standard|premium)$",
            "title": "Tier",
            "type": "string"
          }
        },
        "required": [
          "tier"
        ],
        "title": "QuoteRequest",
        "type": "object"
      },
      "QuoteResult": {
        "description": "A priced quote with its totals and an optional rendered PDF reference.",
        "properties": {
          "grand_total": {
            "$ref": "#/components/schemas/Money"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/QuoteLine"
            },
            "title": "Lines",
            "type": "array"
          },
          "monthly_total": {
            "$ref": "#/components/schemas/Money"
          },
          "one_off_total": {
            "$ref": "#/components/schemas/Money"
          },
          "pdf_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pdf Url"
          },
          "quote_id": {
            "title": "Quote Id",
            "type": "string"
          },
          "tier": {
            "title": "Tier",
            "type": "string"
          },
          "vat_total": {
            "$ref": "#/components/schemas/Money"
          }
        },
        "required": [
          "quote_id",
          "tier",
          "lines",
          "one_off_total",
          "monthly_total",
          "vat_total",
          "grand_total"
        ],
        "title": "QuoteResult",
        "type": "object"
      },
...
      "UtmParams": {
        "additionalProperties": false,
        "description": "Marketing attribution captured with the submission (kept on the CRM lead).",
        "properties": {
          "campaign": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign"
          },
          "content": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "medium": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Medium"
          },
          "source": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "term": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Term"
          }
        },
        "title": "UtmParams",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ViaVitae API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/assessment/": {
      "post": {
        "description": "Record a funnel submission, push a lead+deal to Bitrix24 (UTM kept).\n\nTODO(api): delegate to app.services.quotes_engine.recommend_tier and the\nbitrix24 mapper/sync, persist via the app.api.deps session seam, and return\nthe AssessmentResult. Routers stay thin: validate, call a service, shape.",
        "operationId": "submit_assessment_v1_assessment__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssessmentSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record a completed funnel and return the recommended tier.",
        "tags": [
          "v1",
          "assessment"
        ]
      }
    },
    "/v1/assessment/{assessment_id}": {
      "get": {
        "description": "Fetch a stored assessment scoped to the resolved tenant (RLS).",
        "operationId": "get_assessment_v1_assessment__assessment_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "assessment_id",
            "required": true,
            "schema": {
              "title": "Assessment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessmentResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch a stored assessment for the current tenant.",
        "tags": [
          "v1",
          "assessment"
        ]
      }
    },
...
    "/v1/quotes/": {
      "post": {
        "description": "Price a tier configuration via app.services.quotes_engine.\n\nTODO(api): call quotes_engine.price_quote, render the PDF with\napp.services.pdf (WeasyPrint), push to CRM and email the quote.",
        "operationId": "calculate_quote_v1_quotes__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Price a configuration of products and services.",
        "tags": [
          "v1",
          "quotes"
        ]
      }
    },
...
    "/v1/quotes/{quote_id}/pdf": {
      "get": {
        "description": "Render a stored quote as a signed PDF (WeasyPrint).\n\nTODO(api): load the quote for the tenant and stream the PDF via\napp.services.pdf; sign/number it per the LT invoicing rules.",
        "operationId": "render_quote_pdf_v1_quotes__quote_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "quote_id",
            "required": true,
            "schema": {
              "title": "Quote Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Render Quote Pdf V1 Quotes  Quote Id  Pdf Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Render a stored quote as a signed PDF.",
        "tags": [
          "v1",
          "quotes"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.viavitae.eu/v1"
    }
  ]
}